Replace long with size_t for pointer arithmetic to resolve exceptions.

+++ src/d_fill.c
@@ -61,7 +61,7 @@ void D_FillRect(vrect_t* rect, int color)
 
     dest = ((byte*)vid.buffer + ry * vid.rowbytes + rx);
 
-    if (((rwidth & 0x03) == 0) && (((long)dest & 0x03) == 0)) {
+    if (((rwidth & 0x03) == 0) && (((size_t)dest & 0x03) == 0)) {
         // faster aligned dword clear
         ldest = (unsigned*)dest;
         color += color << 16;
