+++ src/r_draw.c
@@ -347,7 +347,7 @@ void R_EmitCachedEdge(void)
 {
     edge_t* pedge_t;
 
-    pedge_t = (edge_t*)((unsigned long)r_edges + r_pedge->cachededgeoffset);
+    pedge_t = (edge_t*)((size_t)r_edges + r_pedge->cachededgeoffset);
 
     if (!pedge_t->surfs[0]) {
         pedge_t->surfs[0] = surface_p - surfaces;
@@ -425,7 +425,7 @@ void R_RenderFace(msurface_t* fa, int clipflags)
                         continue;
                     }
                 } else {
-                    if ((((unsigned long)edge_p - (unsigned long)r_edges) > r_pedge->cachededgeoffset) && (((edge_t*)((unsigned long)r_edges + r_pedge->cachededgeoffset))->owner == r_pedge)) {
+                    if ((((size_t)edge_p - (size_t)r_edges) > r_pedge->cachededgeoffset) && (((edge_t*)((size_t)r_edges + r_pedge->cachededgeoffset))->owner == r_pedge)) {
                         R_EmitCachedEdge();
                         r_lastvertvalid = false;
                         continue;
@@ -462,7 +462,7 @@ void R_RenderFace(msurface_t* fa, int clipflags)
                 } else {
                     // it's cached if the cached edge is valid and is owned
                     // by this medge_t
-                    if ((((unsigned long)edge_p - (unsigned long)r_edges) > r_pedge->cachededgeoffset) && (((edge_t*)((unsigned long)r_edges + r_pedge->cachededgeoffset))->owner == r_pedge)) {
+                    if ((((size_t)edge_p - (size_t)r_edges) > r_pedge->cachededgeoffset) && (((edge_t*)((size_t)r_edges + r_pedge->cachededgeoffset))->owner == r_pedge)) {
                         R_EmitCachedEdge();
                         r_lastvertvalid = false;
                         continue;
