+++ src/r_surf.c
@@ -557,7 +557,7 @@ void R_DrawSurfaceBlock16(void)
         pbasesource += sourcetstep;
         lightright += lightrightstep;
         lightleft += lightleftstep;
-        prowdest = (unsigned short*)((long)prowdest + surfrowbytes);
+        prowdest = (unsigned short*)((size_t)prowdest + surfrowbytes);
     }
 
     prowdestbase = prowdest;
